requirements.txt
pyspark==3.5.3
beautifulsoup4
pdfplumber
langchain-text-splitters
pandas
pyarrow
python-dotenv
google-generativeai
pinecone
sec-edgar-downloader
